chore: implement vertex adjustments in tracer

+++ src/cpp/vektor/bezier_curve.h
@@ -2,9 +2,9 @@
 #include <glm/vec2.hpp>
 
 struct BezierCurve {
-  glm::vec2 p0, p1, p2, p3;
+  glm::dvec2 p0, p1, p2, p3;
 
-  void apply_scale(float val) {
+  void apply_scale(double val) {
     p0 *= val;
     p1 *= val;
     p2 *= val;
